v0.4.0 - consistent export order + flags (#41)
* Made prints into error.debugs

* [WIP?] Added flags carryover

Scene flags are no longer created from scratch.
If exporting over a file from the same engine, the flags are carried over from the imported file.
Otherwise the flags are taken from the target file.

Not 100% sure if this is a good play - need to study the flags more.

* Changed the compare tool to be order-dependent

Export order may be important for stages, so now we expect import/export order to be consistent.

* Made import/export order-preserving

Stages have two root notes, "instance" and "static".
The games expect "static" to be the first node, but the exporter uses Blender default ordering i.e. alphabetic i.e. "instance" first.
This commit makes the importer remember the sort order in Hierarchy Node data,
and makes the exporter export objects in that order.
This should also fix old-engine hair, which depended on transparent meshes being exported in a specific non-alphabetic order.