Stopped calculating tangents unless possible and necessary
Previously, mesh tangents would always be calculated on export.
Blender requires models to have a UV layer when calculating tangents, so any model without UVs would fail to export here.
Now a recoverable error is reported when importing a mesh that needs calculated tangents but no UVs, and tangents are only calculated for meshes with materials that use them.