Removed broken, unused tail_delta fix code
Older versions had a lot of spare code regarding bone importing, including this part for deciding on the "tail_delta".
Now we just use the "anim_axis", and the "tail_delta" is actually just the world position so the checking code doesn't work anyway.

If bone handling code changes are requested later, we can implement them.
But for now this setup is consistent and hasn't seemed to cause problems.