[WIP] experimented with automatic placing of phys bones
Tried to rework skinned bone gather to detect when a phys bone was moved, and rework the hierarchy in that case.
This relied on assuming all phys bones were in a contiguous segment, which they aren't, so it didn't work.
One *could* try placing padding bones in the old phys indices, and maybe the phys bones would still be displayed? But that's for another day.