Fixed incorrect unfusion solving
Previous unfusion solving was evaluated per fused-vertex V, and only considered the unfusions U[V] while assuming all other fusions between X,Y != V were valid.
This is not the case for disjoint unfusion constraints.
e.g. for fused verts (A,B,A',B'), unfusing A/A' and B/B' would map A -> (A, B, B') which doesn't respect the other constraint.
The new approach is not guaranteed to be minimal but it is guaranteed to respect *all* unfusions.