Moved u8 (-1,1) primitive into structurelib, added tests
Vertex buffer parsing uses the RangeConverterPrimitive class to create a new type,
which remaps a byte [0,255] range to [-1.0, 1.0].
Moving this into structurelib allows it to be tested.
These tests currently fail, which is a good sign that this is the reason the normal-W and tangent-W comparisons in the wider tests fail.