Fix placement of unskinned items in hierarchies
Previously didn't take parentage into account when creating objects.
Now, we create two hierarchies: one in the bones, and one of mesh objects and empties all parented to the armature.
This is a pain in the ass to modify in blender, because all changes need to happen in two places at once,
but we don't need to care because this is "animation mode" import